refactor: move component logic
